# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Support System API",
        "description": "API для системы поддержки",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/support-requests": {
            "get": {
                "tags": [
                    "AdminSupportRequests"
                ],
                "summary": "Получить список всех обращений",
                "operationId": "be4964cc2461a4701eab8feb67120a59",
                "responses": {
                    "200": {
                        "description": "Список всех обращений",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SupportRequest"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/support-requests/{id}/attach-order": {
            "post": {
                "tags": [
                    "AdminSupportRequests"
                ],
                "summary": "Закрепить заказ за обращением",
                "operationId": "f8605d7e897a6b6a5d235d54b5bef201",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID обращения",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заказ успешно закреплен за обращением"
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            }
        },
        "/api/messages/{supportRequestId}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Получить все сообщения по обращению",
                "operationId": "74c6dd1220d4480e1ebb77ca811759f8",
                "parameters": [
                    {
                        "name": "supportRequestId",
                        "in": "path",
                        "description": "ID обращения",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список сообщений",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Message"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Обращение не найдено"
                    }
                }
            }
        },
        "/api/support-requests": {
            "get": {
                "tags": [
                    "SupportRequests"
                ],
                "summary": "Получить список всех обращений",
                "operationId": "b6f7aea71b8ba829545483ab5b4120cf",
                "responses": {
                    "200": {
                        "description": "Список обращений",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SupportRequest"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/support-requests/{id}": {
            "get": {
                "tags": [
                    "SupportRequests"
                ],
                "summary": "Получить конкретное обращение",
                "operationId": "03a8fa157c8014dcf325bd41d276e0a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID обращения",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обращение найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupportRequest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Обращение не найдено"
                    }
                }
            }
        },
        "/api/support-requests/{supportRequestId}/messages": {
            "post": {
                "tags": [
                    "SupportRequests"
                ],
                "summary": "Ответить на обращение",
                "operationId": "703a4642086acb5b916bd78f3e12038b",
                "parameters": [
                    {
                        "name": "supportRequestId",
                        "in": "path",
                        "description": "ID обращения",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Сообщение добавлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/support-requests/{id}/attach-order": {
            "post": {
                "tags": [
                    "SupportRequests"
                ],
                "summary": "Закрепить заказ за обращением",
                "operationId": "286bdb1534351f463130c09d90acc7ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID обращения",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заказ прикреплен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupportRequest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Обращение или заказ не найдены"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Message": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "support_request_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SupportRequest": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "order_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Введите токен в формате \"Bearer {токен}\""
            }
        }
    },
    "tags": [
        {
            "name": "AdminSupportRequests",
            "description": "Операции с обращениями для администраторов"
        },
        {
            "name": "Messages",
            "description": "Операции с сообщениями поддержки"
        },
        {
            "name": "SupportRequests",
            "description": "Операции с обращениями поддержки"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}